// src/doctor/interfaces/qualifications.interface.ts
export interface QualificationsInterface {
  degree: string;
  university: string;
  name_of_hospital_of_practice: string;
  contact_of_hospital_of_practice: string;
  country_of_hospital_of_practice: string;
  state_of_hospital_of_practice: string;
  license_number: string;
  license_expiry_date: string;
}